Switch to fetching phsData.json
Remove redundant code
